Tests/LibWeb/Text/expected/css/unrecognized-font-family-fallback-honors-weight.txt

unrecognized-family bold matches sans-serif bold: true
unrecognized-family bold matches the registered bold face: true
